Fix F1_score
In SelectionPerformanceSingle() the F1 score is calculated as the harmonic mean between precision and recall;

2* precision * recall / (precision + recall)

The F1 score has been set to 1 if precision and recall are 0 i.e., in scenarios where the F1 score is ill-defined.
However, the F1 score should be 0 instead as no true positive is found in these scenarios.